// sariftypesgenerator/src/interfaces.d.ts
// /********************************************************
// *                                                       *
// *   Copyright (C) Microsoft. All rights reserved.       *
// *                                                       *
// ********************************************************/
export interface SarifEnum {
    name: string;
    values: string[];
}